+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### unreleased
 - **upg:** eslint-plugin-import (-> 1.10.0)
 - **upg:** several development dependencies
+- **upg:** several development dependencies
 
 ### [v0.3.0] - 2016-06-27
 - **add:** expose localhost